refactor(fetch): git mv guarded_fetcher.py -> endpoint_guard.py
A pure rename: the file's bytes do not change, so the diff is R100 and there
is nothing in it to review but the new name. Its two importers follow.

The name had been wrong since the transport it was named for was deleted —
what the file holds is the SSRF rule the credential surface runs before it
will persist an object-store endpoint.
